requirements.txt
fastapi==0.110.0
joblib==1.4.2
pydantic==2.10.2
pydantic_core==2.27.1
uvicorn==0.29.0
easyocr==1.7.2
Pillow==9.5.0
numpy==2.0.2
ultralytics==8.3.58
opencv-python==4.10.0